fix(metrics): eliminate unnecessary cast warning under monitoring feature

+++ src/metrics.rs
@@ -566,7 +566,7 @@ pub mod testutil {
                     continue;
                 }
                 if let AggregatedMetrics::F64(MetricData::Histogram(h)) = m.data() {
-                    total += h.data_points().map(|dp| dp.count() as u64).sum::<u64>();
+                    total += h.data_points().map(|dp| dp.count()).sum::<u64>();
                 }
             }
         }
